pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synth-city"
version = "0.1.0"
description = "Agentic layer for building and training competitive models on Bittensor Subnet 50 (Synth)"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.12.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.12.0",
    "arch>=6.2.0",
    "torch>=2.1.0",
    "bittensor>=7.0.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
